add crud medical test

+++ invoice-service/src/main/java/com/kynsoft/invoiceservice/domain/dto/InvoicePaymentDto.java
@@ -17,7 +17,7 @@ public class InvoicePaymentDto {
     private UUID id;
     private String paymentType;
     private String reference;
-    private int plazo;
+    private BigDecimal plazo;
     private String unidadTiempo;
     private BigDecimal amount;
 }
